+++ docs/examples/microwave/microwave.py
@@ -1,9 +1,8 @@
 import tkinter as tk
 from functools import partial
 
-from sismic.interpreter import Interpreter
+from sismic.interpreter import Interpreter, Event
 from sismic.io import import_from_yaml
-from sismic.model import Event
 
 # The two following lines are NOT needed in a typical environment.
 # These lines make sismic available in our testing environment