+++ sismic/code/python.py
@@ -6,10 +6,9 @@ from types import CodeType
 from typing import (Any, Callable, Dict, Iterator, List, Mapping,
                     MutableMapping, cast)
 
+from . import Evaluator
 from ..exceptions import CodeEvaluationError
-from ..model import Event, InternalEvent, Statechart, StateMixin, Transition
-
-from .evaluator import Evaluator
+from ..model import Statechart, StateMixin, Transition, Event, InternalEvent
 
 __all__ = ['PythonEvaluator']