+++ sismic/interpreter/__init__.py
@@ -1,4 +1,5 @@
-from . import helpers
 from .interpreter import Interpreter
+from ..model.events import Event, InternalEvent, MetaEvent
+from ..model.steps import MacroStep, MicroStep
 
-__all__ = ['Interpreter']
+__all__ = ['Interpreter', 'Event', 'InternalEvent', 'MetaEvent', 'MacroStep', 'MicroStep']