+++ sismic/model/__init__.py
@@ -1,7 +1,4 @@
-from .elements import ContractMixin, StateMixin, ActionStateMixin, \
-    TransitionStateMixin, CompositeStateMixin, HistoryStateMixin
-from .elements import BasicState, CompoundState, OrthogonalState, ShallowHistoryState, DeepHistoryState, FinalState
-from .elements import Transition
-from .statechart import Statechart
-from .steps import MicroStep, MacroStep
-from .events import Event, InternalEvent, MetaEvent
+from .elements import *
+from .statechart import *
+from .events import *
+from .steps import *